pengi_service/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
blinker==1.8.2
boto3==1.34.121
botocore==1.34.121
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.6
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.13.3
Flask==3.0.3
frozenlist==1.4.1
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.132.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
grpcio-tools==1.62.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
hyperframe==6.0.1
identify==2.5.35
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.2
langchain-community==0.2.4
langchain-core==0.2.4
langchain-google-genai==1.0.6
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
langsmith==0.1.74
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
nodeenv==1.8.0
numpy==1.26.4
openai==1.31.2
orjson==3.10.3
outcome==1.3.0.post0
packaging==23.2
platformdirs==4.2.0
portalocker==2.8.2
pre-commit==3.7.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.7.3
pydantic-settings==2.3.1
pydantic_core==2.18.4
Pygments==2.18.0
pyinstrument==4.6.2
pyparsing==3.1.2
pypdf==4.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
qdrant-client==1.9.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rsa==4.9
s3transfer==0.10.1
selenium==4.16.0
setuptools==69.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.30
starlette==0.37.2
tenacity==8.3.0
tiktoken==0.7.0
tqdm==4.66.4
trio==0.24.0
trio-websocket==0.11.1
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.30.1
virtualenv==20.25.1
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.3
wsproto==1.2.0
yarl==1.9.4
